+++ langx-java/src/main/java/com/jn/langx/util/datetime/graceful/GracefulDatePrinter.java
@@ -412,7 +412,7 @@ class GracefulDatePrinter implements DatePrinter, Serializable {
         } else if (obj instanceof Calendar) {
             return format((Calendar) obj, toAppendTo);
         } else if (obj instanceof Long) {
-            return format(((Long) obj).longValue(), toAppendTo);
+            return format((Long) obj, toAppendTo);
         } else {
             throw new IllegalArgumentException("Unknown class: " +
                 (obj == null ? "<null>" : obj.getClass().getName()));
